Added Telemetry Service Layer bridge Serial service and Serial parser

+++ TelemetryHub/src/main/java/com/kailas/TelemetryHub/model/TelemetryData.java
@@ -1,6 +1,6 @@
 package com.kailas.TelemetryHub.model;
 
-public record TelemetryData(long Count,
+public record TelemetryData(int Count,
                             int AdcValue,
                             int SamplePeriod){
 }
